requirements.txt
requests==2.31.0
phonenumbers==8.13.27
twilio==8.11.1
python-dotenv==1.0.1